📦📌 Limit Sphinx below v8.2
This is necessary as said version reveals a bug in the
`sphinxcontrib.apidoc` [[1]] extension we've got integrated. It emits
a warning about it and we have configured turning all warnings into
errors.

The commit should be reverted once the upstream fixes the bug and cuts
a new release compatible with newer Sphinx versions.

[1]: https://github.com/sphinx-contrib/apidoc/issues/21